backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "langchain==1.2.7",
    "langchain-openai==1.1.7",
    "pydantic==2.12.5",
    "python-dotenv==1.2.1",
    "sqlmodel>=0.0.31",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
